requirements.txt
wheel>=0.37.1
numpy>=1.23.4
scipy>=1.9.2
pillow>=9.2.0
matplotlib>=3.6.1
ruff>=0.3.7
